+++ zcash_client_backend/src/data_api/wallet.rs
@@ -23,7 +23,8 @@ use {
 use crate::{
     address::RecipientAddress,
     data_api::{
-        error::Error, DecryptedTransaction, SentTransaction, SentTransactionOutput, WalletWrite,
+        error::Error, DecryptedTransaction, PoolType, Recipient, SentTransaction,
+        SentTransactionOutput, WalletWrite,
     },
     decrypt_transaction,
     wallet::OvkPolicy,
@@ -373,14 +374,21 @@ where
     let (tx, tx_metadata) = builder.build(&prover).map_err(Error::Builder)?;
 
     let sent_outputs = request.payments().iter().enumerate().map(|(i, payment)| {
-        let idx = match &payment.recipient_address {
+        let (output_index, recipient) = match &payment.recipient_address {
             // Sapling outputs are shuffled, so we need to look up where the output ended up.
-            // TODO: When we add Orchard support, we will need to trial-decrypt to find them.
-            RecipientAddress::Shielded(_) | RecipientAddress::Unified(_) =>
-                tx_metadata.output_index(i).expect("An output should exist in the transaction for each shielded payment."),
+            RecipientAddress::Shielded(addr) => {
+                let idx = tx_metadata.output_index(i).expect("An output should exist in the transaction for each shielded payment.");
+                (idx, Recipient::Sapling(addr.clone()))
+            }
+            RecipientAddress::Unified(addr) => {
+                // TODO: When we add Orchard support, we will need to trial-decrypt to find them,
+                // and return the appropriate pool type.
+                let idx = tx_metadata.output_index(i).expect("An output should exist in the transaction for each shielded payment.");
+                (idx, Recipient::Unified(addr.clone(), PoolType::Sapling))
+            }
             RecipientAddress::Transparent(addr) => {
                 let script = addr.script();
-                tx.transparent_bundle()
+                let idx = tx.transparent_bundle()
                     .and_then(|b| {
                         b.vout
                             .iter()
@@ -388,13 +396,15 @@ where
                             .find(|(_, tx_out)| tx_out.script_pubkey == script)
                     })
                     .map(|(index, _)| index)
-                    .expect("An output should exist in the transaction for each transparent payment.")
+                    .expect("An output should exist in the transaction for each transparent payment.");
+
+                (idx, Recipient::Transparent(*addr))
             }
         };
 
         SentTransactionOutput {
-            output_index: idx,
-            recipient_address: &payment.recipient_address,
+            output_index,
+            recipient,
             value: payment.amount,
             memo: payment.memo.clone()
         }
@@ -512,12 +522,7 @@ where
 
     // add the sapling output to shield the funds
     builder
-        .add_sapling_output(
-            Some(ovk),
-            shielding_address.clone(),
-            amount_to_shield,
-            memo.clone(),
-        )
+        .add_sapling_output(Some(ovk), shielding_address, amount_to_shield, memo.clone())
         .map_err(Error::Builder)?;
 
     let (tx, tx_metadata) = builder.build(&prover).map_err(Error::Builder)?;
@@ -531,8 +536,8 @@ where
         account,
         outputs: vec![SentTransactionOutput {
             output_index,
-            recipient_address: &RecipientAddress::Shielded(shielding_address),
             value: amount_to_shield,
+            recipient: Recipient::InternalAccount(account, PoolType::Sapling),
             memo: Some(memo.clone()),
         }],
         fee_amount: fee,