+++ zcash_client_backend/src/lib.rs
@@ -19,4 +19,4 @@ pub mod wallet;
 pub mod welding_rig;
 pub mod zip321;
 
-pub use decrypt::{decrypt_transaction, DecryptedOutput};
+pub use decrypt::{decrypt_transaction, DecryptedOutput, TransferType};